backend/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
cryptography==38.0.1
defusedxml==0.7.1
Django==3.2.6
django-cors-headers==3.13.0
django-extra-fields==3.0.2
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.2
pep8-naming==0.13.2
Pillow==9.2.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.6.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.9.0